Add dbus_replaylocks.py
This service reacts on "Replaying" dbus messages of the dbus2vdr plugin for vdr.
If a recording on a file system which does not equal the VDRDIR filesystem is played, the index file of this recording is opened until replay has been stopped. This prevents unwanted umounts by autofs inacitivty timeouts, if the replay is paused.